+++ Marlin/src/HAL/SAMD51/eeprom_flash.cpp
@@ -64,7 +64,7 @@ bool PersistentStore::access_finish() {
   return true;
 }
 
-bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, uint16_t *crc) {
+bool PersistentStore::write_data(uint16_t &pos, const uint8_t *value, size_t size, uint16_t *crc) {
   while (size--) {
     const uint8_t v = *value;
     SYNC(NVMCTRL->SEESTAT.bit.BUSY);
@@ -79,7 +79,7 @@ bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, ui
   return false;
 }
 
-bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
+bool PersistentStore::read_data(uint16_t &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
   while (size--) {
     SYNC(NVMCTRL->SEESTAT.bit.BUSY);
     uint8_t c = ((volatile uint8_t *)SEEPROM_ADDR)[pos];
